Filedict: Move get_file_dict
