+++ coalib/processes/Processing.py
@@ -21,8 +21,7 @@ from coalib.results.RESULT_SEVERITY import RESULT_SEVERITY
 from coalib.results.SourceRange import SourceRange
 from coalib.settings.Setting import glob_list
 from coalib.parsing.Globbing import fnmatch
-from coalib.files.Fileproxy import Fileproxy
-
+from coalib.files.Filedict import get_file_dict
 
 ACTIONS = [ApplyPatchAction,
            PrintDebugMessageAction,
@@ -237,35 +236,6 @@ def print_result(results,
                   file_dict,
                   file_diff_dict)
     return retval or len(results) > 0, patched_results
-
-
-def get_file_dict(filename_list, log_printer):
-    """
-    Reads all files into a dictionary.
-
-    :param filename_list: List of names of paths to files to get contents of.
-    :param log_printer:   The logger which logs errors.
-    :return:              Reads the content of each file into a dictionary
-                          with filenames as keys.
-    """
-    file_dict = {}
-    for filename in filename_list:
-        try:
-            file_dict[filename] = Fileproxy(filename)
-        except UnicodeDecodeError:
-            log_printer.warn("Failed to read file '{}'. It seems to contain "
-                             "non-unicode characters. Leaving it "
-                             "out.".format(filename))
-        except OSError as exception:  # pragma: no cover
-            log_printer.log_exception("Failed to read file '{}' because of "
-                                      "an unknown error. Leaving it "
-                                      "out.".format(filename),
-                                      exception,
-                                      log_level=LOG_LEVEL.WARNING)
-
-    log_printer.debug("Files that will be checked:\n" +
-                      "\n".join(file_dict.keys()))
-    return file_dict
 
 
 def filter_raising_callables(it, exception, *args, **kwargs):
