+++ tests/files/FiledictTestFiles/testcode.c
@@ -1,0 +1,1 @@
+#include <stdio.h>
